submission-research2/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
setuptools==75.6.0
sqlparse==0.5.2
uritemplate==4.1.1
